Switched from graphqlgen to graphql-code-generator.
