--- meteor-app/imports/api/graphql/.gitignore
@@ -1,1 +1,0 @@
-graphqlgen.ts
