--- meteor-app/imports/api/graphql/graphqlgen.yml
@@ -1,17 +1,0 @@
-language: typescript
-
-schema: ./schema.graphql
-context: ./resolvers.ts:Context
-models:
-  files:
-    - ../models/comment/comment.ts
-    - ../models/misc.ts
-    - ../models/company/company.ts
-    - ../models/job-ad/job-ad.ts
-    - ../models/location/location.ts
-    - ../models/review/review.ts
-    - ../models/salary/salary.ts
-    - ../models/user/user.ts
-    - ../models/vote/vote.ts
-
-output: ./graphqlgen.ts
