+++ meteor-app/imports/api/graphql/index.ts
@@ -1,4 +1,4 @@
 import typeDefs from "./schema.graphql";
-import resolvers from "./resolvers.js";
+import { resolvers } from "./resolvers";
 
 export { typeDefs, resolvers };
