+++ meteor-app/imports/api/graphql/resolvers/.gitignore
@@ -1,0 +1,1 @@
+resolvers-types.ts
