+++ meteor-app/imports/api/graphql/resolvers/context.ts
@@ -1,0 +1,5 @@
+import * as dataModel from "imports/api/models";
+
+export type Context = {
+	user: dataModel.User;
+};
