+++ meteor-app/imports/api/graphql/resolvers/index.ts
@@ -1,0 +1,1 @@
+export * from "./resolvers";
