+++ meteor-app/imports/api/graphql/resolvers/mutation.ts
@@ -1,0 +1,12 @@
+import * as dataModel from "imports/api/models";
+
+import { MutationResolvers, RewardStatus } from "./resolvers-types";
+
+export const Mutation: MutationResolvers.Resolvers = {
+	claimWroteAReview: (_obj, args, context, _info) =>
+		dataModel.claimWroteAReview(
+			context.user,
+			args.phoneNumber,
+			args.paymentMethod
+		) as Promise<RewardStatus>,
+};
